+++ api/src/main/java/io/strimzi/api/kafka/model/kafka/KafkaStatus.java
@@ -6,10 +6,12 @@ package io.strimzi.api.kafka.model.kafka;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import io.strimzi.api.annotations.DeprecatedProperty;
 import io.strimzi.api.kafka.model.common.Constants;
 import io.strimzi.api.kafka.model.kafka.cruisecontrol.KafkaAutoRebalanceStatus;
 import io.strimzi.api.kafka.model.kafka.listener.ListenerStatus;
 import io.strimzi.crdgenerator.annotations.Description;
+import io.strimzi.crdgenerator.annotations.PresentInVersions;
 import io.sundr.builder.annotations.Buildable;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
@@ -58,6 +60,9 @@ public class KafkaStatus extends Status {
         this.kafkaNodePools = kafkaNodePools;
     }
 
+    @Deprecated
+    @DeprecatedProperty(description = "The `registeredNodeIds` property is deprecated and it is not used anymore. It will be removed in the future.")
+    @PresentInVersions("v1beta2")
     @Description("Registered node IDs used by this Kafka cluster. " +
             "This field is used for internal purposes only and will be removed in the future.")
     public List<Integer> getRegisteredNodeIds() {