+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/operator/assembly/KafkaNodeUnregistrationTest.java
@@ -12,9 +12,11 @@ import io.vertx.junit5.Checkpoint;
 import io.vertx.junit5.VertxExtension;
 import io.vertx.junit5.VertxTestContext;
 import org.apache.kafka.clients.admin.Admin;
+import org.apache.kafka.clients.admin.DescribeClusterOptions;
+import org.apache.kafka.clients.admin.DescribeClusterResult;
 import org.apache.kafka.clients.admin.UnregisterBrokerResult;
 import org.apache.kafka.common.KafkaFuture;
-import org.apache.kafka.common.errors.BrokerIdNotRegisteredException;
+import org.apache.kafka.common.Node;
 import org.apache.kafka.common.errors.TimeoutException;
 import org.apache.kafka.common.internals.KafkaFutureImpl;
 import org.junit.jupiter.api.AfterAll;
@@ -24,6 +26,9 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
 
 import java.util.List;
+import java.util.Set;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 
 import static org.hamcrest.CoreMatchers.hasItems;
 import static org.hamcrest.CoreMatchers.is;
@@ -57,38 +62,7 @@ public class KafkaNodeUnregistrationTest {
         AdminClientProvider mockProvider = ResourceUtils.adminClientProvider(mockAdmin);
 
         Checkpoint async = context.checkpoint();
-        KafkaNodeUnregistration.unregisterNodes(Reconciliation.DUMMY_RECONCILIATION, vertx, mockProvider, null, null, List.of(1874, 1919))
-                .onComplete(context.succeeding(v -> context.verify(() -> {
-                    assertThat(unregisteredNodeIdCaptor.getAllValues().size(), is(2));
-                    assertThat(unregisteredNodeIdCaptor.getAllValues(), hasItems(1874, 1919));
-
-                    async.flag();
-                })));
-    }
-
-    @Test
-    void testUnknownNodeUnregistration(VertxTestContext context) {
-        Admin mockAdmin = ResourceUtils.adminClient();
-
-        ArgumentCaptor<Integer> unregisteredNodeIdCaptor = ArgumentCaptor.forClass(Integer.class);
-        when(mockAdmin.unregisterBroker(unregisteredNodeIdCaptor.capture())).thenAnswer(i -> {
-            if (i.getArgument(0, Integer.class) == 1919)   {
-                KafkaFutureImpl<Void> unregistrationFuture = new KafkaFutureImpl<>();
-                unregistrationFuture.completeExceptionally(new BrokerIdNotRegisteredException("Unknown node"));
-                UnregisterBrokerResult ubr = mock(UnregisterBrokerResult.class);
-                when(ubr.all()).thenReturn(unregistrationFuture);
-                return ubr;
-            } else {
-                UnregisterBrokerResult ubr = mock(UnregisterBrokerResult.class);
-                when(ubr.all()).thenReturn(KafkaFuture.completedFuture(null));
-                return ubr;
-            }
-        });
-
-        AdminClientProvider mockProvider = ResourceUtils.adminClientProvider(mockAdmin);
-
-        Checkpoint async = context.checkpoint();
-        KafkaNodeUnregistration.unregisterNodes(Reconciliation.DUMMY_RECONCILIATION, vertx, mockProvider, null, null, List.of(1874, 1919))
+        KafkaNodeUnregistration.unregisterBrokerNodes(Reconciliation.DUMMY_RECONCILIATION, vertx, mockProvider, null, null, Set.of(1874, 1919))
                 .onComplete(context.succeeding(v -> context.verify(() -> {
                     assertThat(unregisteredNodeIdCaptor.getAllValues().size(), is(2));
                     assertThat(unregisteredNodeIdCaptor.getAllValues(), hasItems(1874, 1919));
@@ -119,12 +93,58 @@ public class KafkaNodeUnregistrationTest {
         AdminClientProvider mockProvider = ResourceUtils.adminClientProvider(mockAdmin);
 
         Checkpoint async = context.checkpoint();
-        KafkaNodeUnregistration.unregisterNodes(Reconciliation.DUMMY_RECONCILIATION, vertx, mockProvider, null, null, List.of(1874, 1919))
+        KafkaNodeUnregistration.unregisterBrokerNodes(Reconciliation.DUMMY_RECONCILIATION, vertx, mockProvider, null, null, Set.of(1874, 1919))
                 .onComplete(context.failing(v -> context.verify(() -> {
                     assertThat(unregisteredNodeIdCaptor.getAllValues().size(), is(2));
                     assertThat(unregisteredNodeIdCaptor.getAllValues(), hasItems(1874, 1919));
 
                     async.flag();
                 })));
+    }
+
+    @Test
+    void testListRegisteredBrokerNodes(VertxTestContext context) throws InterruptedException {
+        Admin mockAdmin = ResourceUtils.adminClient();
+
+        List<Node> cluster = List.of(
+                new Node(0, "my-broker-0", 9092, "rack-1", true),
+                new Node(1, "my-broker-1", 9092, "rack-1", true),
+                new Node(2, "my-broker-2", 9092, "rack-2", false),
+                new Node(3, "my-broker-3", 9092, "rack-2", false),
+                new Node(4, "my-broker-4", 9092, "rack-2", false)
+        );
+
+        ArgumentCaptor<DescribeClusterOptions> dcoCaptor = ArgumentCaptor.forClass(DescribeClusterOptions.class);
+        when(mockAdmin.describeCluster(dcoCaptor.capture())).thenAnswer(i -> {
+            DescribeClusterResult dcr = mock(DescribeClusterResult.class);
+            if (i.getArgument(0, DescribeClusterOptions.class).includeFencedBrokers()) {
+                when(dcr.nodes()).thenReturn(KafkaFuture.completedFuture(cluster));
+            } else {
+                when(dcr.nodes()).thenReturn(KafkaFuture.completedFuture(cluster.stream().filter(node -> !node.isFenced()).toList()));
+            }
+            return dcr;
+        });
+
+        AdminClientProvider mockProvider = ResourceUtils.adminClientProvider(mockAdmin);
+
+        CountDownLatch latch = new CountDownLatch(2);
+
+        KafkaNodeUnregistration.listRegisteredBrokerNodes(Reconciliation.DUMMY_RECONCILIATION, vertx, mockProvider, null, null, true)
+                .onComplete(context.succeeding(nodes -> context.verify(() -> {
+                    assertThat(nodes.size(), is(5));
+                    latch.countDown();
+                })));
+
+        KafkaNodeUnregistration.listRegisteredBrokerNodes(Reconciliation.DUMMY_RECONCILIATION, vertx, mockProvider, null, null, false)
+                .onComplete(context.succeeding(nodes -> context.verify(() -> {
+                    assertThat(nodes.size(), is(3));
+                    for (Node node : nodes) {
+                        assertThat(node.isFenced(), is(false));
+                    }
+                    latch.countDown();
+                })));
+
+        latch.await(10, TimeUnit.SECONDS);
+        context.completeNow();
     }
 }